Rename ProfileView to UserProfileView
